01-python-skript-pro-llm-api/pyproject.toml
[project]
name = "3-function-calls"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "curl-cffi>=0.11.4",
    "openai>=1.66.3",
    "python-dotenv>=1.0.1",
    "yfinance>=0.2.54",
]
